Took out old rendering formats.
Decreases size of executable, which is good, at the expense of a few old
replays we won't be able to read any more through the main version.
I've copied the old executable, however, to allow  the occasional
browsing of old games as one might wish, but we'll overall move away
from that.


Former-commit-id: c83fb1a29ab3dc05a519fd8b845fb47ecf82968c [formerly c38a410602c900065483cf29e0d87753b592910b]
Former-commit-id: 82ab1990c8992b8d6901d93c96f5a3f08dbb2c73
Former-commit-id: 823df967dc6749689a7ab2830017098ddffba1a3