Fix occasional crash in Halite environment (fixes #23)
The game environment was terminating sporadically due to an
uncaught std::string thrown in handle_frame_networking. The culprit
was read_trailing_input, which caught BotInputError, but failed
to account for the fact that get_string can also throw std::string.
read_trailing_input gives get_string a timeout of 0 milliseconds,
so if it took any non-negligible amount of time to read the string,
it would throw an error instead of returning normally, and since
read_trailing_input was called inside a catch() {} block, the
catchall handler would not catch the exception.