Fix damage calculation with multiple targets
When multiple targets were involved, the engine was incorrectly
setting the cooldown on the weapon after applying damage to the
first target, causing other ships to not be attacked. The fix
here takes advantage of the 2-phase calculation of damage, and
uses the fact that we first compute targets and damage before
applying it. Now, we apply damage based on the attack event
information that we were already tracking, so we know the attack
must have happened - we do not have to recheck the cooldown or
other matters.

First spotted by @MoreGames on the Halite II forums.