Increase SERIALIZATION_PRECISION to 6
I don't see any real downside to this. This number only determines the precision sent to the clients, it doesn't affect replays.

We could even go higher e.g. 8 or 10. But I suppose it's nice to be able to look at the comms visually if one wants to understand the format (e.g. [I did so here](https://forums.halite.io/t/halite-comms-protocol-explained/411/2)). So perhaps 6 is fine.