Increase pipe buffer size on Linux
The default pipe buffer size is 65536 bytes on Linux. If there are
a lot of ships, the serialized map state might be larger than that,
so we will encounter an error when trying to write it, and thus we
will eject all bots. On Linux, we can resize the buffer to avoid
this problem. This sets the size of the buffer to the maximum
size allowable (which is 1 MiB on my system) on Linux.

The alternative is to write the map in pieces, waiting for the
bot to read and clear the buffer before writing more. However,
then we either have to count the time to write the map as part of
the bot's turn time, or we have to introduce a separate timeout.
Otherwise, a bot could intentionally let its pipe fill up to buy
extra processing time while the engine waits for it to read the
map. And if we add a separate timeout, I fear bots will have to
implement this pathological behavior, because it's a way to
maximize the amount of time you get per turn.