Set last match time for challenges to prevent getting stuck
If a challenge game could not be played for some reason (e.g. one of
the participants has no bots), this would stop all challenge games,
because the matchmaker would keep trying to issue games for the same
challenge.  This changes it so that even if a game could not be
played, the most recent match time is still updated, so that we
don't get stuck on that challenge.